Replace old table with new
