+++ components/styled/table/src/molecules/table.tsx
@@ -1,19 +1,17 @@
-import { FC, useState, useEffect } from 'react'
+import { FC, useState } from 'react'
 import {
-  ColumnDef,
-  createColumnHelper,
   flexRender,
   getCoreRowModel,
   useReactTable,
-  ExpandedState,
   getExpandedRowModel,
   getSortedRowModel,
+  ExpandedState,
   SortingState,
-  SortDirection,
 } from '@tanstack/react-table'
-import uuid from 'react-uuid'
 import styled from '@emotion/styled'
 import { CSS_RESET, TRANSLUCENT_BRIGHT_BLUE_TABLE, TRANSLUCENT_MID_GREY, SCROLLBAR } from '@ltht-react/styles'
+import { CellProps } from './table-cell'
+import createColumns from './table-methods'
 
 const Container = styled.div`
   ${CSS_RESET};
@@ -57,94 +55,13 @@ const StyledTableData = styled.td`
   }
 `
 
-const columnHelper = createColumnHelper<DataRow>()
-
-const generateColumnsFromHeadersRecursively = (headers?: Header[]): ColumnDef<DataRow, CellData | unknown>[] =>
-  headers?.map((header) =>
-    header.subheaders
-      ? columnHelper.group({
-          id: header.id ?? uuid(),
-          header: () => header.header,
-          columns: generateColumnsFromHeadersRecursively(header.subheaders),
-        })
-      : (columnHelper.accessor(header.accessor, {
-          id: header.accessor,
-          cell: (info) => {
-            const stringValue = (info?.getValue() as string) ?? ''
-            return header.cell
-              ? header.cell({ value: stringValue, columnId: info.column.id, row: info.row.original })
-              : stringValue
-          },
-          header: () => header.header,
-        }) as ColumnDef<DataRow, CellData | unknown>)
-  ) ?? []
-
-const generateRowsFromCellRows = (cellRows: CellRow[]): DataRow[] =>
-  cellRows.map((cellRow) => {
-    const mappedCell: DataRow = { subRows: [] }
-    const mappedCellRender: Record<string, FC<ICellProps>> = {}
-
-    cellRow.cells.forEach((cell) => {
-      mappedCell[cell.key] = cell.value
-      if (cell.render) {
-        mappedCellRender[cell.key] = cell.render
-      }
-    })
-
-    return {
-      ...mappedCell,
-      ...{
-        rowId: cellRow.id ?? '',
-        render: cellRow.render ? cellRow.render : (props: ICellProps) => <>{props.value}</>,
-        renderCells: mappedCellRender,
-      },
-      subRows: generateRowsFromCellRows(cellRow.subRows ?? []),
-    }
-  })
-
-const getExpanderColumn = (): ColumnDef<DataRow, CellData | unknown> =>
-  columnHelper.accessor('expander', {
-    header: ({ table }) => (
-      <span
-        title="Toggle All Rows Expanded"
-        onClick={table.getToggleAllRowsExpandedHandler()}
-        style={{ cursor: 'pointer' }}
-      >
-        {table.getIsAllRowsExpanded() ? '▲' : '►'}
-      </span>
-    ),
-    cell: ({ row }) =>
-      row.getCanExpand() ? (
-        <span
-          {...{
-            onClick: row.getToggleExpandedHandler(),
-            style: { cursor: 'pointer', paddingLeft: `${row.depth * 2}rem` },
-          }}
-        >
-          {row.getIsExpanded() ? '▲' : '►'}
-        </span>
-      ) : null,
-  }) as ColumnDef<DataRow, CellData | unknown>
-
-export default function Table({ tableData }: IProps): JSX.Element {
-  const [columns, setColumns] = useState<ColumnDef<DataRow, CellData | unknown>[]>([])
-  const [data, setData] = useState<DataRow[]>([])
+const Table: FC<IProps> = ({ tableData }) => {
   const [expanded, setExpanded] = useState<ExpandedState>({})
   const [sorting, setSorting] = useState<SortingState>([])
 
-  useEffect(() => {
-    const columnArray = generateColumnsFromHeadersRecursively(tableData.headers)
-    const dataArray = generateRowsFromCellRows(tableData.rows)
-
-    setColumns(
-      dataArray.some((x: DataRow) => x.subRows.length > 0) ? [getExpanderColumn(), ...columnArray] : columnArray
-    )
-    setData(dataArray)
-  }, [tableData])
-
   const table = useReactTable({
-    data,
-    columns,
+    data: tableData.rows,
+    columns: createColumns(tableData),
     state: {
       expanded,
       sorting,
@@ -156,12 +73,6 @@ export default function Table({ tableData }: IProps): JSX.Element {
     getExpandedRowModel: getExpandedRowModel(),
     getSortedRowModel: getSortedRowModel(),
   })
-
-  const getSortIcon = (sortDirection: SortDirection): string | null =>
-    ({
-      asc: ' 🔼',
-      desc: ' 🔽',
-    }[sortDirection] ?? null)
 
   return (
     <Container>
@@ -186,7 +97,6 @@ export default function Table({ tableData }: IProps): JSX.Element {
                     }}
                   >
                     {header.isPlaceholder ? null : flexRender(header.column.columnDef.header, header.getContext())}
-                    {getSortIcon(header.column.getIsSorted() as SortDirection)}
                   </StyledTableHeader>
                 )
               )}
@@ -215,44 +125,24 @@ export default function Table({ tableData }: IProps): JSX.Element {
   )
 }
 
-interface IProps {
-  tableData: TableData
-}
-
-export interface ICellProps {
-  value: string
-  row: DataRow
-  columnId: string
+export type DataEntity = Record<string, CellProps | DataEntity[]> & {
+  subRows?: DataEntity[]
 }
 
 export interface Header {
-  id?: string
-  header: string | JSX.Element
-  accessor: string
-  subheaders?: Header[]
-  cell?: FC<ICellProps>
-}
-
-export interface Cell {
-  key: string
-  value: string | JSX.Element
-  render?: FC<ICellProps>
-}
-
-export interface CellRow {
-  id?: string
-  cells: Cell[]
-  render?: FC<ICellProps>
-  subRows?: CellRow[]
+  type: 'accessor' | 'group' | 'display'
+  id: string
+  cellProps: CellProps
+  subHeaders?: Header[]
 }
 
 export interface TableData {
   headers: Header[]
-  rows: CellRow[]
+  rows: DataEntity[]
 }
 
-export declare type CellData = string | FC<ICellProps> | Record<string, FC<ICellProps>> | DataRow[] | JSX.Element
-export declare type DataRow = {
-  [key: string]: CellData
-  subRows: DataRow[]
+interface IProps {
+  tableData: TableData
 }
+
+export default Table
